# Optmize `interop.flow.StreamSubscriber.onNext`