Since the `reactive-streams` spec mentions that all calls must be done _"serially"_ these changes optimize the tight loop of multiple consequential `onNext` calls.

-----

## Benchmark results

### Before the optimization

| Benchmark | (fibers) | (iterations) | Mode | Cnt | Score | Error | Units |
| - | - | - | - | - | - | - | - |
| FlowInteropBenchmark.fastPublisher | 1000 |   1024 | thrpt | 20 | 25.938 | ± 0.306 | ops/s |
| FlowInteropBenchmark.fastPublisher | 1000 |   5120 | thrpt | 20 |  6.519 | ± 0.015 | ops/s |
| FlowInteropBenchmark.fastPublisher | 1000 |  10240 | thrpt | 20 |  3.201 | ± 0.168 | ops/s |
| FlowInteropBenchmark.fastPublisher | 1000 |  51200 | thrpt | 20 |  0.680 | ± 0.013 | ops/s |
| FlowInteropBenchmark.fastPublisher | 1000 | 512000 | thrpt | 20 |  0.062 | ± 0.001 | ops/s |

### After the optimization

| Benchmark | (fibers) | (iterations) | Mode | Cnt | Score | Error | Units |
| - | - | - | - | - | - | - | - |
| FlowInteropBenchmark.fastPublisher | 1000 |   1024 | thrpt | 20 | 63.846 | ± 0.094 | ops/s |
| FlowInteropBenchmark.fastPublisher | 1000 |   5120 | thrpt | 20 | 22.314 | ± 0.107 | ops/s |
| FlowInteropBenchmark.fastPublisher | 1000 |  10240 | thrpt | 20 | 12.662 | ± 0.028 | ops/s |
| FlowInteropBenchmark.fastPublisher | 1000 |  51200 | thrpt | 20 |  2.736 | ± 0.012 | ops/s |
| FlowInteropBenchmark.fastPublisher | 1000 | 512000 | thrpt | 20 |  0.275 | ± 0.002 | ops/s |

-----

## Analysis

For only `2` chunks of `512` elements the improvements were almost `2.5` times better.
For `20` chunks, the improvements were almost `4` times better.
But the improvements for `1000` chunks were just `4.4` times better.

It seems then that in a fully sequential and CPU-bound `Publisher`, the new `Subscriber` is roughly 4 times more efficient. As long as the chunk size is adequate, and there are enough chunks to flatten the overhead of the general machinery.
Overall, I would say this optimization looks great, even if realistic use cases won't get such a big increase, it is likely the effects will be noticeable.

-----

> Lies, Damn Lies, and Benchmarks